pyproject.toml
[project]
name = "dealagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "openai>=1.90.0",
    "playwright>=1.53.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "tavily-python>=0.7.9",
    "uvicorn[standard]>=0.35.0",
    "vllm>=0.9.2",
]
